+++ Hearthstone Deck Tracker/Controls/Overlay/Constructed/RelatedCardsPanel/RelatedCardsPanelViewModel.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Input;
@@ -57,6 +57,7 @@ public class RelatedCardsPanelViewModel : ViewModel
 		}
 	}
 
+	[LocalizedProp]
 	public string HeaderText => CardName != null
 		? string.Format(LocUtil.Get("TheOutfinder_Label_PoolHeader"), CardName, Cards?.Count ?? 0)
 		: LocUtil.Get("TheOutfinder_Label_DefaultHeader");